Cover more syntax.
The objective to begin with is to get 100% coverage for Scala syntax
and never (!!) to change break the semantics of the formatted code,
even if this comes at the cost of less beautiful output. To begin
with, we only optimize for correctness and once we have full confidence
that

- formatted(code).structure == code.structure, and
- no comment every gets lost

then we can start polishing the output. While polishing the output,
then we must never regress on these two main objectives.