Steal TreeSyntax trick for handling wrapping parens.
So far we have taken an ad-hoc approach to guessing when parentheses are
not sufficient. This does not scale, it's way more tricky than infix
operators, it's nearly part of every relationship between a tree node
and its children. Early results look very promising, and require minimal
changes to the printer. Reasoning about missing parens bugs is a lot
simpler now.